requirements.txt
pyocr
pillow
opencv-contrib-python
argparse